myFund/Models/CommandHandler.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using System.Windows.Input;

namespace myFund.Models
{
    public class CommandHandler<T> : ICommand
    {
        private Action<T> _action;
        private Func<T, bool> _canExecute;
        public CommandHandler(Action<T> action, Func<T,bool> canExecute)
        {
            _action = action;
            _canExecute = canExecute;
        }

        public bool CanExecute(object parameter)
        {
            return _canExecute((T)parameter);
        }

        public event EventHandler CanExecuteChanged;

        public void Execute(object parameter)
        {
            _action((T)parameter);
        }
    }
}
